Refactored collect_dofs() to a new Dofs() class. Created finite-differencing function
