+++ simsopt/__init__.py
@@ -2,7 +2,9 @@ from .util import *
 from .optimizable import *
 from .surface import *
 from .functions import *
+from .dofs import *
 from .least_squares_problem import *
+from .finite_difference import *
 from .vmec import *
 
 #all = ['Parameter']
